Organize and document hackathon solutions for portfolio
This change completely restructures the repository to create a polished portfolio piece from the hackathon solutions.

Key changes include:
- A new, professional directory structure with each solution in its own folder.
- A detailed main README.md that introduces the project, highlights the hackathon achievement, and provides a table of contents for all solutions.
- Individual README.md files for each of the 10 tasks, with inferred problem descriptions and detailed explanations of the solution approaches.
- Added comments to the source code of all solutions to improve readability and explain the logic.
- A .gitignore file to exclude unnecessary compiled files and caches.
- Renamed the "Task question" directory to "task-questions" for consistency.
- Corrected a typo in the filename for task H.
