+++ .gitignore
@@ -1,0 +1,8 @@
+# C++ compiled files
+*.out
+*.exe
+solution
+
+# Python cache
+__pycache__/
+*.pyc
