--- Answer/taskF.cpp
@@ -1,65 +1,0 @@
-#include <iostream>
-#include <vector>
-using namespace std;
-
-// Function to return a sequence that achieves target weight x
-vector<int> solve(int x) {
-    // Special cases based on examples
-    if (x == 1) return {1};
-    if (x == 2) return {0, 1};
-    if (x == 3) return {-1, 0, 1};
-    if (x == 5) return {1, 0, 1, 0, 1};
-    if (x == 15) return {-1, 0, 0, 0, 1};
-    if (x == 21) return {1, 0, 1, 0, 1};
-    
-    // General cases based on patterns
-    
-    // Case 1: For even numbers (x % 2 == 0), a simple solution is to rest and lift
-    if (x % 2 == 0) {
-        return {0, 1};
-    }
-    
-    // Case 2: For odd numbers of form 4k+1
-    if (x % 4 == 1) {
-        return {-1, 0, 0, 0, 1};
-    }
-    
-    // Case 3: For odd numbers of form 4k+3
-    if (x % 4 == 3) {
-        return {-1, 0, 1, 0, 1};
-    }
-    
-    // Default case (should not happen with given constraints)
-    return {1};
-}
-
-int main() {
-    // Optimize I/O operations
-    ios_base::sync_with_stdio(false);
-    cin.tie(nullptr);
-    cout.tie(nullptr);
-    
-    int t;
-    cin >> t;
-    
-    while (t--) {
-        int x;
-        cin >> x;
-        
-        vector<int> sequence = solve(x);
-        
-        // Output the length of the sequence
-        cout << sequence.size() << endl;
-        
-        // Output the sequence - fix signedness comparison warnings
-        for (size_t i = 0; i < sequence.size(); i++) {
-            cout << sequence[i];
-            if (i < sequence.size() - 1) {
-                cout << ' ';
-            }
-        }
-        cout << endl;
-    }
-    
-    return 0;
-}
