+++ README.md
@@ -1,2 +1,57 @@
-# Code-Gym-2025-Q-A
-Code Gym 2025. Solution and answer to the task which helped me get third place in the hackathon. 
+# Code Gym 2025 Hackathon: 3rd Place Finish
+
+This repository showcases the solutions that secured a 3rd place finish at the Code Gym 2025 hackathon. This project is a testament to my ability to solve complex problems under pressure and write efficient code in a competitive environment.
+
+## About the Hackathon
+
+The Code Gym 2025 was an intense 24-hour coding competition where participants were challenged with a series of algorithmic problems. My solutions were written in C++ and Python, demonstrating my versatility in different programming languages.
+
+## Solutions
+
+Here is a breakdown of the tasks and my solutions. Each solution includes a `README.md` file with a more detailed explanation of the problem and my approach.
+
+| Task | Language | Solution | Problem Description |
+| :--- | :--- | :--- | :--- |
+| **Task A** | C++ | [View Solution](./solutions/task-a/solution.cpp) | [View Problem](./task-questions/Task%20A.pdf) |
+| **Task B** | Python | [View Solution](./solutions/task-b/solution.py) | [View Problem](./task-questions/Task%20B.pdf) |
+| **Task C** | C++ | [View Solution](./solutions/task-c/solution.cpp) | [View Problem](./task-questions/Task%20C.pdf) |
+| **Task D** | C++ | [View Solution](./solutions/task-d/solution.cpp) | [View Problem](./task-questions/Task%20D.pdf) |
+| **Task E** | C++ | [View Solution](./solutions/task-e/solution.cpp) | [View Problem](./task-questions/Task%20E.pdf) |
+| **Task F** | C++ | [View Solution](./solutions/task-f/solution.cpp) | [View Problem](./task-questions/Task%20F.pdf) |
+| **Task G** | Python | [View Solution](./solutions/task-g/solution.py) | [View Problem](./task-questions/Task%20G.pdf) |
+| **Task H** | Python | [View Solution](./solutions/task-h/solution.py) | [View Problem](./task-questions/Task%20H.pdf) |
+| **Task I** | C++ | [View Solution](./solutions/task-i/solution.cpp) | [View Problem](./task-questions/Task%20I.pdf) |
+| **Task J** | C++ | [View Solution](./solutions/task-j/solution.cpp) | [View Problem](./task-questions/Task%20J.pdf) |
+
+## How to Run the Solutions
+
+### C++
+
+To compile and run the C++ solutions, you will need a C++ compiler (like g++).
+
+```bash
+# Navigate to the solution directory
+cd solutions/task-a
+
+# Compile the solution
+g++ -o solution solution.cpp
+
+# Run the solution
+./solution
+```
+
+### Python
+
+To run the Python solutions, you will need Python 3 installed.
+
+```bash
+# Navigate to the solution directory
+cd solutions/task-b
+
+# Run the solution
+python solution.py
+```
+
+## Acknowledgements
+
+I would like to thank the organizers of Code Gym 2025 for putting on such a great event. This experience was incredibly valuable and has further fueled my passion for coding.
