+++ solutions/task-c/solution.cpp
