+++ solutions/task-e/README.md
@@ -1,0 +1,32 @@
+# Task E: Conditional XOR Sum
+
+## Problem Description
+
+*This problem description is inferred from the solution code.*
+
+You are given `t` test cases. Each test case consists of a range of integers `[l, r]` and two additional integers, `i` and `k`.
+
+The goal is to calculate the XOR sum of all integers `x` in the range `[l, r]` (inclusive) that do **not** satisfy the condition `x % (2^i) == k`.
+
+## Solution Approach
+
+The solution leverages mathematical properties of the XOR operation to avoid a naive iteration, which would be too slow for large ranges.
+
+1.  **XOR Sum of a Range:** The core idea is to first calculate the XOR sum of all numbers in the range `[l, r]`, and then "remove" the XOR contribution of the numbers that should be excluded.
+    *   The XOR sum of numbers from 1 to `n`, let's call it `xorUpTo(n)`, can be calculated in O(1) time based on the value of `n mod 4`. The code implements this in the `xorUpTo` function.
+    *   The XOR sum of a range `[l, r]` can then be calculated as `xorUpTo(r) ^ xorUpTo(l - 1)`. This is a standard and highly efficient technique.
+
+2.  **Identifying Excluded Numbers:** The numbers to be excluded are those `x` in `[l, r]` such that `x % (2^i) == k`. These numbers form an arithmetic progression with a common difference of `2^i`.
+    *   The solution finds the first number `start` in the range `[l, r]` that satisfies this condition.
+    *   It then calculates how many such numbers (`count`) exist in the range.
+
+3.  **Removing the Contribution of Excluded Numbers:**
+    *   The XOR sum of an arithmetic progression where the common difference is a power of two has special properties. However, the code uses a simpler and clever observation.
+    *   Let the set of excluded numbers be `S = {s_1, s_2, ..., s_m}`. We need to calculate `(XOR sum of S)`.
+    *   The code observes that if the count of these numbers is even, their total XOR sum is 0 (this is not generally true, but seems to be a simplification that works for this problem's constraints, possibly because the XOR sum of pairs cancels out in a specific way).
+    *   If the count is odd, the total XOR sum is just one of the numbers (`start`). This is a strong assumption and likely relies on specific properties of the problem constraints not immediately obvious from the code alone.
+    *   The final result is obtained by taking the total XOR sum of the range and XORing it with the calculated XOR sum of the excluded numbers.
+
+4.  **Edge Cases:** The code handles the special case where `i = 0`. In this case, `2^i = 1`, and `x % 1` is always 0. So, if `k=0`, all numbers are excluded (XOR sum is 0). If `k != 0`, no numbers are excluded.
+
+This approach is very efficient, with a time complexity of O(1) per test case, making it suitable for large ranges and a high number of queries.
