+++ solutions/task-f/README.md
@@ -1,0 +1,21 @@
+# Task F: Constructive Sequence
+
+## Problem Description
+
+*This problem description is inferred from the solution code. The exact rules for how the sequence generates the target value `x` are not immediately clear, but the solution is based on pattern recognition.*
+
+The task is to construct a sequence of integers (`1`, `0`, or `-1`) for a given target integer `x`. The solution must produce a valid sequence for any given `x`.
+
+## Solution Approach
+
+The solution is a classic example of a constructive algorithm based on pattern matching and case analysis. Instead of a single mathematical formula, it uses a series of checks to find a pre-computed or pattern-based solution for the input `x`.
+
+1.  **Hardcoded Special Cases:** The code begins by handling several specific values of `x` (1, 2, 3, 5, 15, 21) with hardcoded sequences. This is a common technique to handle tricky edge cases or base cases for a pattern.
+
+2.  **General Pattern for Even Numbers:** For any even number `x`, the solution returns the sequence `{0, 1}`. This suggests that this sequence is a valid construction for all even numbers under the problem's (unclear) rules.
+
+3.  **General Pattern for Odd Numbers:** For odd numbers, the solution further divides them into two cases based on their value modulo 4.
+    *   If `x % 4 == 1` (e.g., 1, 5, 9, ...), the sequence `{-1, 0, 0, 0, 1}` is returned.
+    *   If `x % 4 == 3` (e.g., 3, 7, 11, ...), the sequence `{-1, 0, 1, 0, 1}` is returned.
+
+This case-based approach covers all possible integer inputs for `x`. The solution is very fast as it only involves a few simple checks. The main challenge in creating such a solution is to find these patterns in the first place, which often involves a combination of mathematical reasoning, observation, and testing on small examples.
