+++ solutions/task-g/solution.py
