+++ solutions/task-j/solution.cpp
